[DE85] Rebuild thread need to be exited when sender thread exits (#96)
