+++ lib/libzrepl/data_conn.c
@@ -33,6 +33,7 @@
 #include <uzfs_io.h>
 #include <uzfs_rebuilding.h>
 #include <zrepl_mgmt.h>
+#include <uzfs_mgmt.h>
 #include "mgmt_conn.h"
 #include "data_conn.h"
 
@@ -341,6 +342,53 @@ drop_refcount:
 	uzfs_zinfo_drop_refcnt(zinfo);
 }
 
+static void
+uzfs_zvol_append_to_fd_list(zvol_info_t *zinfo, int fd)
+{
+	zinfo_fd_t *new_zinfo_fd = kmem_alloc(sizeof (zinfo_fd_t), KM_SLEEP);
+	new_zinfo_fd->fd = fd;
+
+	(void) pthread_mutex_lock(&zinfo->zinfo_mutex);
+#ifdef DEBUG
+	zinfo_fd_t *zinfo_fd = NULL;
+	STAILQ_FOREACH(zinfo_fd, &zinfo->fd_list, fd_link) {
+		if (zinfo_fd->fd == fd) {
+			ASSERT(1 == 0);
+		}
+	}
+#endif
+	STAILQ_INSERT_TAIL(&zinfo->fd_list, new_zinfo_fd, fd_link);
+	LOG_DEBUG("Appending fd %d for zvol %s", fd, zinfo->name);
+	(void) pthread_mutex_unlock(&zinfo->zinfo_mutex);
+}
+
+static void
+uzfs_zvol_remove_from_fd_list(zvol_info_t *zinfo, int fd)
+{
+	zinfo_fd_t *zinfo_fd = NULL;
+
+	(void) pthread_mutex_lock(&zinfo->zinfo_mutex);
+#ifdef DEBUG
+	int count = 0;
+	STAILQ_FOREACH(zinfo_fd, &zinfo->fd_list, fd_link) {
+		if (zinfo_fd->fd == fd)
+			count++;
+	}
+	ASSERT(count == 1);
+#endif
+	zinfo_fd = STAILQ_FIRST(&zinfo->fd_list);
+	while (zinfo_fd != NULL) {
+		if (zinfo_fd->fd == fd) {
+			STAILQ_REMOVE(&zinfo->fd_list, zinfo_fd,
+			    zinfo_fd_s, fd_link);
+			kmem_free(zinfo_fd, sizeof (zinfo_fd_t));
+			break;
+		}
+		zinfo_fd = STAILQ_NEXT(zinfo_fd, fd_link);
+	}
+	(void) pthread_mutex_unlock(&zinfo->zinfo_mutex);
+}
+
 void
 uzfs_zvol_rebuild_dw_replica(void *arg)
 {
@@ -359,6 +407,8 @@ uzfs_zvol_rebuild_dw_replica(void *arg)
 
 	sfd = rebuild_args->fd;
 	zinfo = rebuild_args->zinfo;
+
+	uzfs_zvol_append_to_fd_list(zinfo, sfd);
 
 	if ((rc = setsockopt(sfd, SOL_SOCKET, SO_LINGER, &lo, sizeof (lo)))
 	    != 0) {
@@ -496,6 +546,8 @@ next_step:
 	}
 
 exit:
+	uzfs_zvol_remove_from_fd_list(zinfo, sfd);
+
 	mutex_enter(&zinfo->zv->rebuild_mtx);
 	if (rc != 0) {
 		uzfs_zvol_set_rebuild_status(zinfo->zv,
@@ -854,7 +906,9 @@ uzfs_zvol_rebuild_scanner_callback(off_t offset, size_t len,
 	hdr.len = len;
 	hdr.flags = ZVOL_OP_FLAG_REBUILD;
 	hdr.status = ZVOL_OP_STATUS_OK;
-	if (zinfo->state == ZVOL_INFO_STATE_OFFLINE)
+
+	if ((zinfo->state == ZVOL_INFO_STATE_OFFLINE) ||
+	    (zinfo->is_io_ack_sender_created == B_FALSE))
 		return (-1);
 
 	LOG_DEBUG("IO number for rebuild %ld", metadata->io_num);
@@ -883,13 +937,13 @@ uzfs_zvol_rebuild_scanner(void *arg)
 	zvol_info_t	*zinfo = NULL;
 	zvol_io_hdr_t	hdr;
 	int 		rc = 0;
-	zvol_rebuild_t  warg;
+	zvol_rebuild_t	warg;
 	char 		*name;
 	blk_metadata_t	metadata;
 	uint64_t	rebuild_req_offset;
 	uint64_t	rebuild_req_len;
 	zvol_io_cmd_t	*zio_cmd;
-	struct linger lo = { 1, 0 };
+	struct linger	lo = { 1, 0 };
 
 	if ((rc = setsockopt(fd, SOL_SOCKET, SO_LINGER, &lo, sizeof (lo)))
 	    != 0) {
@@ -897,16 +951,23 @@ uzfs_zvol_rebuild_scanner(void *arg)
 		goto exit;
 	}
 read_socket:
+	if ((zinfo != NULL) &&
+	    ((zinfo->state == ZVOL_INFO_STATE_OFFLINE) ||
+	    (zinfo->is_io_ack_sender_created == B_FALSE)))
+		goto exit;
+
 	rc = uzfs_zvol_read_header(fd, &hdr);
-	if ((rc != 0) || ((zinfo != NULL) &&
-	    (zinfo->state == ZVOL_INFO_STATE_OFFLINE)))
+	if ((rc != 0) ||
+	    ((zinfo != NULL) &&
+	    ((zinfo->state == ZVOL_INFO_STATE_OFFLINE) ||
+	    (zinfo->is_io_ack_sender_created == B_FALSE))))
 		goto exit;
 
 	LOG_DEBUG("op_code=%d io_seq=%ld", hdr.opcode, hdr.io_seq);
 
 	/* Handshake yet to happen */
 	if ((hdr.opcode != ZVOL_OPCODE_HANDSHAKE) && (zinfo == NULL)) {
-		LOG_DEBUG("Wrong opcode:%d, expecting handshake\n", hdr.opcode);
+		LOG_DEBUG("Wrong opcode:%d, expecting handshake", hdr.opcode);
 		rc = -1;
 		goto exit;
 	}
@@ -939,6 +1000,9 @@ read_socket:
 			}
 
 			LOG_INFO("Rebuild scanner started on zvol %s", name);
+
+			uzfs_zvol_append_to_fd_list(zinfo, fd);
+
 			kmem_free(name, hdr.len);
 			warg.zinfo = zinfo;
 			warg.fd = fd;
@@ -954,7 +1018,11 @@ read_socket:
 			    "Rebuild Req offset: %ld, Rebuild Req length: %ld",
 			    metadata.io_num, rebuild_req_offset,
 			    rebuild_req_len);
-
+#if DEBUG
+			if (inject_error.delay.helping_replica_rebuild_step
+			    == 1)
+				sleep(5);
+#endif
 			rc = uzfs_get_io_diff(zinfo->zv, &metadata,
 			    uzfs_zvol_rebuild_scanner_callback,
 			    rebuild_req_offset, rebuild_req_len, &warg);
@@ -989,6 +1057,8 @@ exit:
 	if (zinfo != NULL) {
 		LOG_INFO("Closing rebuild connection for zvol %s", zinfo->name);
 		remove_pending_cmds_to_ack(fd, zinfo);
+		uzfs_zvol_remove_from_fd_list(zinfo, fd);
+
 		uzfs_zinfo_drop_refcnt(zinfo);
 	} else {
 		LOG_INFO("Closing rebuild connection");
@@ -996,5 +1066,469 @@ exit:
 
 	shutdown(fd, SHUT_RDWR);
 	close(fd);
+	zk_thread_exit();
+}
+
+/*
+ * (Re)Initializes zv's state variables.
+ * This fn need to be called to use zv across network disconnections.
+ * Lock protection and life of zv need to be managed by caller
+ */
+static void
+reinitialize_zv_state(zvol_state_t *zv)
+{
+	if (zv == NULL)
+		return;
+	zv->zv_metavolblocksize = 0;
+
+	uzfs_zvol_set_status(zv, ZVOL_STATUS_DEGRADED);
+	uzfs_zvol_set_rebuild_status(zv, ZVOL_REBUILDING_INIT);
+}
+
+/*
+ * This func takes care of sending potentially multiple read blocks each
+ * prefixed by metainfo.
+ */
+static int
+uzfs_send_reads(int fd, zvol_io_cmd_t *zio_cmd)
+{
+	zvol_io_hdr_t 	*hdr = &zio_cmd->hdr;
+	struct zvol_io_rw_hdr read_hdr;
+	metadata_desc_t	*md;
+	size_t	rel_offset = 0;
+	int	rc = 0;
+
+	/* special case for missing metadata */
+	if (zio_cmd->metadata_desc == NULL) {
+		read_hdr.io_num = 0;
+		/*
+		 * read_hdr.len should be adjusted back
+		 * to actual read request size now
+		 */
+		read_hdr.len = hdr->len -
+		    sizeof (struct zvol_io_rw_hdr);
+		rc = uzfs_zvol_socket_write(fd, (char *)&read_hdr,
+		    sizeof (read_hdr));
+		if (rc != 0)
+			return (rc);
+		/* Data that need to be sent is equal to read_hdr.len */
+		rc = uzfs_zvol_socket_write(fd, zio_cmd->buf, read_hdr.len);
+		return (rc);
+	}
+
+	/*
+	 * TODO: Optimize performance by combining multiple writes to a single
+	 * system call either by copying all data to larger buffer or using
+	 * vector write.
+	 */
+	for (md = zio_cmd->metadata_desc; md != NULL; md = md->next) {
+		read_hdr.io_num = md->metadata.io_num;
+		read_hdr.len = md->len;
+		rc = uzfs_zvol_socket_write(fd, (char *)&read_hdr,
+		    sizeof (read_hdr));
+		if (rc != 0)
+			goto end;
+
+		rc = uzfs_zvol_socket_write(fd,
+		    (char *)zio_cmd->buf + rel_offset, md->len);
+		if (rc != 0)
+			goto end;
+		rel_offset += md->len;
+	}
+
+end:
+	FREE_METADATA_LIST(zio_cmd->metadata_desc);
+	zio_cmd->metadata_desc = NULL;
+
+	return (rc);
+}
+
+/*
+ * One thread per LUN/vol. This thread works
+ * on queue and it sends ack back to client on
+ * a given fd.
+ * There are two types of clients - one is iscsi target, and,
+ * other is a replica which undergoes rebuild.
+ * Need to exit from thread when there are network errors
+ * on fd related to iscsi target.
+ */
+static void
+uzfs_zvol_io_ack_sender(void *arg)
+{
+	int fd;
+	int md_len;
+	zvol_info_t		*zinfo;
+	thread_args_t 		*thrd_arg;
+	zvol_io_cmd_t 		*zio_cmd = NULL;
+
+	thrd_arg = (thread_args_t *)arg;
+	fd = thrd_arg->fd;
+	zinfo = thrd_arg->zinfo;
+	kmem_free(arg, sizeof (thread_args_t));
+
+	prctl(PR_SET_NAME, "ack_sender", 0, 0, 0);
+
+	LOG_INFO("Started ack sender for zvol %s fd: %d", zinfo->name, fd);
+
+	while (1) {
+		int rc = 0;
+		(void) pthread_mutex_lock(&zinfo->zinfo_mutex);
+		zinfo->zio_cmd_in_ack = NULL;
+		while (1) {
+			if ((zinfo->state == ZVOL_INFO_STATE_OFFLINE) ||
+			    (zinfo->conn_closed == B_TRUE)) {
+				(void) pthread_mutex_unlock(
+				    &zinfo->zinfo_mutex);
+				goto exit;
+			}
+			if (STAILQ_EMPTY(&zinfo->complete_queue)) {
+				zinfo->io_ack_waiting = 1;
+				pthread_cond_wait(&zinfo->io_ack_cond,
+				    &zinfo->zinfo_mutex);
+				zinfo->io_ack_waiting = 0;
+			}
+			else
+				break;
+		}
+
+		zio_cmd = STAILQ_FIRST(&zinfo->complete_queue);
+		STAILQ_REMOVE_HEAD(&zinfo->complete_queue, cmd_link);
+		zinfo->zio_cmd_in_ack = zio_cmd;
+		(void) pthread_mutex_unlock(&zinfo->zinfo_mutex);
+
+		LOG_DEBUG("ACK for op: %d, seq-id: %ld",
+		    zio_cmd->hdr.opcode, zio_cmd->hdr.io_seq);
+
+		/* account for space taken by metadata headers */
+		if (zio_cmd->hdr.status == ZVOL_OP_STATUS_OK &&
+		    zio_cmd->hdr.opcode == ZVOL_OPCODE_READ) {
+			md_len = 0;
+			for (metadata_desc_t *md = zio_cmd->metadata_desc;
+			    md != NULL;
+			    md = md->next) {
+				md_len++;
+			}
+			/* we need at least one header even if no metadata */
+			if (md_len == 0)
+				md_len++;
+			zio_cmd->hdr.len += (md_len *
+			    sizeof (struct zvol_io_rw_hdr));
+		}
+
+		rc = uzfs_zvol_socket_write(zio_cmd->conn,
+		    (char *)&zio_cmd->hdr, sizeof (zio_cmd->hdr));
+		if (rc == -1) {
+			LOG_ERRNO("socket write err");
+			zinfo->zio_cmd_in_ack = NULL;
+			/*
+			 * exit due to network errors on fd related
+			 * to iscsi target
+			 */
+			if (zio_cmd->conn == fd) {
+				zio_cmd_free(&zio_cmd);
+				goto exit;
+			}
+			zio_cmd_free(&zio_cmd);
+			continue;
+		}
+
+		if (zio_cmd->hdr.opcode == ZVOL_OPCODE_READ) {
+			if (zio_cmd->hdr.status == ZVOL_OP_STATUS_OK) {
+				/* Send data read from disk */
+				rc = uzfs_send_reads(zio_cmd->conn, zio_cmd);
+				if (rc == -1) {
+					zinfo->zio_cmd_in_ack = NULL;
+					LOG_ERRNO("socket write err");
+					if (zio_cmd->conn == fd) {
+						zio_cmd_free(&zio_cmd);
+						goto exit;
+					}
+				}
+			}
+			atomic_inc_64(&zinfo->read_req_ack_cnt);
+		} else {
+			if (zio_cmd->hdr.opcode == ZVOL_OPCODE_WRITE)
+				atomic_inc_64(&zinfo->write_req_ack_cnt);
+			else if (zio_cmd->hdr.opcode == ZVOL_OPCODE_SYNC)
+				atomic_inc_64(&zinfo->sync_req_ack_cnt);
+		}
+		zinfo->zio_cmd_in_ack = NULL;
+		zio_cmd_free(&zio_cmd);
+	}
+exit:
+	zinfo->zio_cmd_in_ack = NULL;
+	shutdown(fd, SHUT_RDWR);
+	LOG_INFO("Data connection for zvol %s closed on fd: %d",
+	    zinfo->name, fd);
+
+	remove_pending_cmds_to_ack(fd, zinfo);
+
+	(void) pthread_mutex_lock(&zinfo->zinfo_mutex);
+	zinfo->conn_closed = B_FALSE;
+	zinfo->is_io_ack_sender_created = B_FALSE;
+	(void) pthread_mutex_unlock(&zinfo->zinfo_mutex);
+
+	uzfs_zinfo_drop_refcnt(zinfo);
+
+	zk_thread_exit();
+}
+/*
+ * Process open request on data connection, the first message.
+ *
+ * Return status meaning:
+ *   != 0: OPEN failed, stop reading data from connection.
+ *   == 0 && zinfopp == NULL: OPEN failed, recoverable error
+ *   == 0 && zinfopp != NULL: OPEN succeeded, proceed with other commands
+ */
+static int
+open_zvol(int fd, zvol_info_t **zinfopp)
+{
+	int		rc;
+	zvol_io_hdr_t	hdr;
+	zvol_op_open_data_t open_data;
+	zvol_info_t	*zinfo = NULL;
+	zvol_state_t	*zv = NULL;
+	kthread_t	*thrd_info;
+	thread_args_t 	*thrd_arg;
+	int		rele_dataset_on_error = 0;
+
+	/*
+	 * If we don't know the version yet, be more careful when
+	 * reading header
+	 */
+	if (uzfs_zvol_read_header(fd, &hdr) != 0) {
+		LOG_ERR("error reading open header");
+		return (-1);
+	}
+	if (hdr.opcode != ZVOL_OPCODE_OPEN) {
+		LOG_ERR("zvol must be opened first");
+		return (-1);
+	}
+	if (hdr.len != sizeof (open_data)) {
+		LOG_ERR("Invalid payload length for open");
+		return (-1);
+	}
+	rc = uzfs_zvol_socket_read(fd, (char *)&open_data, sizeof (open_data));
+	if (rc != 0) {
+		LOG_ERR("Payload read failed");
+		return (-1);
+	}
+
+	open_data.volname[MAX_NAME_LEN - 1] = '\0';
+	zinfo = uzfs_zinfo_lookup(open_data.volname);
+	if (zinfo == NULL) {
+		LOG_ERR("zvol %s not found", open_data.volname);
+		hdr.status = ZVOL_OP_STATUS_FAILED;
+		goto open_reply;
+	}
+	if (zinfo->state != ZVOL_INFO_STATE_ONLINE) {
+		LOG_ERR("zvol %s is not online", open_data.volname);
+		hdr.status = ZVOL_OP_STATUS_FAILED;
+		goto open_reply;
+	}
+	if (zinfo->is_io_ack_sender_created != B_FALSE) {
+		LOG_ERR("zvol %s ack sender already present",
+		    open_data.volname);
+		hdr.status = ZVOL_OP_STATUS_FAILED;
+		goto open_reply;
+	}
+
+	zv = zinfo->zv;
+	ASSERT3P(zv, !=, NULL);
+
+	if (zv->zv_metavolblocksize != 0) {
+		LOG_ERR("there might be already a data connection for %s",
+		    open_data.volname);
+		hdr.status = ZVOL_OP_STATUS_FAILED;
+		goto open_reply;
+	}
+
+	ASSERT3P(zv->zv_status, ==, ZVOL_STATUS_DEGRADED);
+	ASSERT3P(zv->rebuild_info.zv_rebuild_status, ==, ZVOL_REBUILDING_INIT);
+
+	if ((zv->zv_status != ZVOL_STATUS_DEGRADED) ||
+	    ((zv->rebuild_info.zv_rebuild_status != ZVOL_REBUILDING_INIT) &&
+	    (zv->rebuild_info.zv_rebuild_status != ZVOL_REBUILDING_FAILED))) {
+		LOG_ERR("as status for %s is %d or rebuild status is %d",
+		    open_data.volname, zv->zv_status,
+		    zv->rebuild_info.zv_rebuild_status);
+		hdr.status = ZVOL_OP_STATUS_FAILED;
+		goto open_reply;
+	}
+	// validate block size (only one bit is set in the number)
+	if (open_data.tgt_block_size == 0 ||
+	    (open_data.tgt_block_size & (open_data.tgt_block_size - 1)) != 0) {
+		LOG_ERR("Invalid block size");
+		hdr.status = ZVOL_OP_STATUS_FAILED;
+		goto open_reply;
+	}
+
+	(void) pthread_mutex_lock(&zinfo->zinfo_mutex);
+	/*
+	 * Hold objset if this is the first query for the zvol. This can happen
+	 * in case that the target creates data connection directly without
+	 * getting the endpoint through mgmt connection first.
+	 */
+	rele_dataset_on_error = 0;
+	if (zv->zv_objset == NULL) {
+		if (uzfs_hold_dataset(zv) != 0) {
+			(void) pthread_mutex_unlock(&zinfo->zinfo_mutex);
+			LOG_ERR("Failed to hold zvol during open");
+			hdr.status = ZVOL_OP_STATUS_FAILED;
+			goto open_reply;
+		}
+		rele_dataset_on_error = 1;
+	}
+	if (uzfs_update_metadata_granularity(zv,
+	    open_data.tgt_block_size) != 0) {
+		(void) pthread_mutex_unlock(&zinfo->zinfo_mutex);
+		if (rele_dataset_on_error == 1)
+			uzfs_rele_dataset(zv);
+		LOG_ERR("Failed to set granularity of metadata");
+		hdr.status = ZVOL_OP_STATUS_FAILED;
+		goto open_reply;
+	}
+	/*
+	 * TODO: Once we support multiple concurrent data connections for a
+	 * single zvol, we should probably check that the timeout is the same
+	 * for all data connections.
+	 */
+	uzfs_update_ionum_interval(zinfo, open_data.timeout);
+	zinfo->timeout = open_data.timeout;
+	*zinfopp = zinfo;
+
+	zinfo->conn_closed = B_FALSE;
+	zinfo->is_io_ack_sender_created = B_TRUE;
+	(void) pthread_mutex_unlock(&zinfo->zinfo_mutex);
+	thrd_arg = kmem_alloc(sizeof (thread_args_t), KM_SLEEP);
+	thrd_arg->fd = fd;
+	thrd_arg->zinfo = zinfo;
+	uzfs_zinfo_take_refcnt(zinfo);
+	thrd_info = zk_thread_create(NULL, 0,
+	    (thread_func_t)uzfs_zvol_io_ack_sender, (void *)thrd_arg, 0, NULL,
+	    TS_RUN, 0, PTHREAD_CREATE_DETACHED);
+	VERIFY3P(thrd_info, !=, NULL);
+
+	hdr.status = ZVOL_OP_STATUS_OK;
+
+open_reply:
+	hdr.len = 0;
+	rc = uzfs_zvol_socket_write(fd, (char *)&hdr, sizeof (hdr));
+
+	/*
+	 * Reinitializing zv states during this error is taken care
+	 * in open_zvol caller
+	 */
+	if (rc == -1)
+		LOG_ERR("Failed to send reply for open request");
+	if (hdr.status != ZVOL_OP_STATUS_OK) {
+		ASSERT3P(*zinfopp, ==, NULL);
+		if (zinfo != NULL)
+			uzfs_zinfo_drop_refcnt(zinfo);
+		return (-1);
+	}
+	return (rc);
+}
+
+/*
+ * IO-Receiver would be per ZVOL, it would be
+ * responsible for receiving IOs on given socket.
+ */
+void
+uzfs_zvol_io_receiver(void *arg)
+{
+	int		rc;
+	int		fd = (uintptr_t)arg;
+	zvol_info_t	*zinfo = NULL;
+	zvol_io_cmd_t	*zio_cmd;
+	zvol_io_hdr_t	hdr;
+
+	prctl(PR_SET_NAME, "io_receiver", 0, 0, 0);
+
+	/* First command should be OPEN */
+	while (zinfo == NULL) {
+		if (open_zvol(fd, &zinfo) != 0) {
+			if ((zinfo != NULL) &&
+			    (zinfo->is_io_ack_sender_created))
+				goto exit;
+			shutdown(fd, SHUT_RDWR);
+			goto thread_exit;
+		}
+	}
+
+	LOG_INFO("Data connection associated with zvol %s fd: %d",
+	    zinfo->name, fd);
+
+	while ((rc = uzfs_zvol_socket_read(fd, (char *)&hdr, sizeof (hdr))) ==
+	    0) {
+		if ((zinfo->state == ZVOL_INFO_STATE_OFFLINE))
+			break;
+
+		if (hdr.opcode != ZVOL_OPCODE_WRITE &&
+		    hdr.opcode != ZVOL_OPCODE_READ &&
+		    hdr.opcode != ZVOL_OPCODE_SYNC) {
+			LOG_ERR("Unexpected opcode %d", hdr.opcode);
+			break;
+		}
+
+		if (((hdr.opcode == ZVOL_OPCODE_WRITE) ||
+		    (hdr.opcode == ZVOL_OPCODE_READ)) && !hdr.len) {
+			LOG_ERR("Zero Payload size for opcode %d", hdr.opcode);
+			break;
+		} else if ((hdr.opcode == ZVOL_OPCODE_SYNC) && hdr.len > 0) {
+			LOG_ERR("Unexpected payload for opcode %d", hdr.opcode);
+			break;
+		}
+
+		zio_cmd = zio_cmd_alloc(&hdr, fd);
+		/* Read payload for commands which have it */
+		if (hdr.opcode == ZVOL_OPCODE_WRITE) {
+			rc = uzfs_zvol_socket_read(fd, zio_cmd->buf, hdr.len);
+			if (rc != 0) {
+				zio_cmd_free(&zio_cmd);
+				break;
+			}
+		}
+
+		if (zinfo->state == ZVOL_INFO_STATE_OFFLINE) {
+			zio_cmd_free(&zio_cmd);
+			break;
+		}
+		/* Take refcount for uzfs_zvol_worker to work on it */
+		uzfs_zinfo_take_refcnt(zinfo);
+		zio_cmd->zv = zinfo;
+		taskq_dispatch(zinfo->uzfs_zvol_taskq, uzfs_zvol_worker,
+		    zio_cmd, TQ_SLEEP);
+	}
+exit:
+	(void) pthread_mutex_lock(&zinfo->zinfo_mutex);
+	zinfo->conn_closed = B_TRUE;
+	/*
+	 * Send signal to ack sender so that it can free
+	 * zio_cmd, close fd and exit.
+	 */
+	if (zinfo->io_ack_waiting) {
+		rc = pthread_cond_signal(&zinfo->io_ack_cond);
+	}
+	/*
+	 * wait for ack thread to exit to avoid races with new
+	 * connections for the same zinfo
+	 */
+	while (zinfo->conn_closed && zinfo->is_io_ack_sender_created) {
+		(void) pthread_mutex_unlock(&zinfo->zinfo_mutex);
+		usleep(1000);
+		(void) pthread_mutex_lock(&zinfo->zinfo_mutex);
+	}
+	(void) pthread_mutex_unlock(&zinfo->zinfo_mutex);
+
+	shutdown_fds_related_to_zinfo(zinfo);
+
+	zinfo->io_ack_waiting = 0;
+
+	reinitialize_zv_state(zinfo->zv);
+	uzfs_zinfo_drop_refcnt(zinfo);
+thread_exit:
+	close(fd);
+	LOG_INFO("Data connection closed on fd: %d", fd);
 	zk_thread_exit();
 }
